Rename subset.data.frame to fast_subset.data.frame
This for the primary purpose to prevent the masking
warning to show up on loading the package. We think
that this is justified as the time savings from the
faster subset method are probably limited. We decided
to keep the function within the package under a different
name in case the need to use it arises in the future.